requirements.txt
appnope==0.1.4
APScheduler==3.11.2
asttokens==3.0.1
build==1.4.2
certifi==2026.2.25
charset-normalizer==3.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
fonttools==4.62.1
idna==3.11
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
loguru==0.7.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
narwhals==2.18.1
nest-asyncio==1.6.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.20.0
pyparsing==3.3.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyzmq==27.1.0
requests==2.33.0
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.48
stack-data==0.6.3
pmdarima==2.0.4
TA-Lib==0.6.8
tornado==6.5.5
traitlets==5.14.3
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
wcwidth==0.6.0